Crate vault_rs

Vault Client

This is a thin wrapper around the HTTP API for Vault.

Modules

secrets
Implementation of the Various Vault Secret Engines

sys
Implements API wrappers for the System Backend endpoints

Structs

Authentication
Authentication data from Vault

Client
Vault API Client

LeasedData
Wrapped Vault Secret with Lease Data

Method
The Request Method (VERB)

ResponseData
Vault Generic Response Data

Secret
A wrapper around a String with custom implementation of Display and Debug to not leak secrets during logging.

Enums

Error
Error type for this library

Response
Generic Vault Response

TokenType
Type of token from Vault See Vault Documentation

Traits

Vault
Trait implementing the basic API operations for Vault