Struct vault_rs::Client

pub struct Client { /* fields omitted */ }

Vault API Client

Methods

impl Client

pub fn new<S1, S2, S3>(
vault_address: Option<S1>,
vault_token: Option<S2>,
root_ca: Option<S3>,
revoke_self_on_drop: bool
) -> Result<Self, Error> where
S1: AsRef<str>,
S2: AsRef<str>,
S3: AsRef<str>,

Create a new Vault Client.

You can optionally provide a reqwest::Client if you have specific needs like custom root CA certificate or require client authentication

pub fn from_environment<S1, S2, S3>(
address: Option<S1>,
token: Option<S2>,
ca_cert: Option<S3>
) -> Result<Self, Error> where
S1: AsRef<str>,
S2: AsRef<str>,
S3: AsRef<str>,

Create a client from environment variables. You can provide alternative sources of the parameters with the optional arguments

The vnrionment variables are:

VAULT_ADDR: Vault Address

VAULT_TOKEN: Vault Token

VAULT_CACERT: Path to the CA Certificate for Vault

pub fn address(&self) -> &str

Returns the Vault address

pub async fn revoke_self<'_>(&'_ self) -> Result<(), Error>

Revoke the Vault token itself

If successful, the Vault Token can no longer be used