Trait vault_rs::Vault

pub trait Vault {
    fn read(&self, path: &str, method: Method) -> Result<Response, Error>;

    fn read_with_query<T: Serialize + ?Sized>(
        &self, 
        path: &str, 
        method: Method, 
        query: &T
    ) -> Result<Response, Error>;

    fn write<T: Serialize>(
        &self, 
        path: &str, 
        payload: &T, 
        method: Method, 
        response_expected: bool
    ) -> Result<Response, Error>;

    fn get(&self, path: &str) -> Result<Response, Error> { ... }

    fn get_with_query<T: Serialize + ?Sized>(
        &self, 
        path: &str, 
        query: &T
    ) -> Result<Response, Error> { ... }

    fn list(&self, path: &str) -> Result<Response, Error> { ... }

    fn post<T: Serialize>(
        &self, 
        path: &str, 
        payload: &T, 
        response_expected: bool
    ) -> Result<Response, Error> { ... }

    fn put<T: Serialize>(
        &self, 
        path: &str, 
        payload: &T, 
        response_expected: bool
    ) -> Result<Response, Error> { ... }

    fn delete(
        &self, 
        path: &str, 
        response_expected: bool
    ) -> Result<Response, Error> { ... }
}

Trait implementing the basic API operations for Vault

Required methods

fn read(&self, path: &str, method: Method) -> Result<Response, Error>

Read a generic Path from Vault

fn read_with_query<T: Serialize + ?Sized>(
&self,
path: &str,
method: Method,
query: &T
) -> Result<Response, Error>

Read a generic Path from Vault with Query

fn write<T: Serialize>(
&self,
path: &str,
payload: &T,
method: Method,
response_expected: bool
) -> Result<Response, Error>

Write to a generic Path in Vault.

Loading content...
Provided methods

fn get(&self, path: &str) -> Result<Response, Error>

Convenience method to Get a generic path from Vault

fn get_with_query<T: Serialize + ?Sized>(
&self,
path: &str,
query: &T
) -> Result<Response, Error>

Convenience method to Get a generic path from Vault

fn list(&self, path: &str) -> Result<Response, Error>

Convenience method to List a generic path from Vault

fn post<T: Serialize>(
&self,
path: &str,
payload: &T,
response_expected: bool
) -> Result<Response, Error>

Convenience method to Post to a generic path to Vault

fn put<T: Serialize>(
&self,
path: &str,
payload: &T,
response_expected: bool
) -> Result<Response, Error>

Convenience method to Put to a generic path to Vault

fn delete(&self, path: &str, response_expected: bool) -> Result<Response, Error>

Convenience method to Delete a Path from Vault

Loading content...